requirements.txt
celery==4.4.0rc4
PyYAML==5.1.2
requests
lxml